+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project closely adheres to [Semantic Versioning](https://semver.org/spe
 - Add - debug and browser logs to steps
 - Change - use pathlib for paths
 - Add - metadata to features, scenarios and steps
+- Add - CONFIG now can save multiple snapshots in a stack
 
 ## 1.2.4
 - Fix - restore state after tab info